+++ src/menu.js
@@ -1,3 +1,5 @@
+let lastUsedPath = null; // global variable
+
 const Menus = {
   IDS: {
     TABSTRIP: {
@@ -5,7 +7,8 @@ const Menus = {
       TO_RIGHT: "save-in-_-_-SI-to-right",
       TO_RIGHT_MATCH: "save-in-_-_-SI-to-right-match",
       OPENED_FROM_TAB: "save-in-_-_-SI-opened-from-tab"
-    }
+    },
+    ROOT: "save-in-_-_-root"
   },
 
   makeSeparator: (() => {
@@ -16,13 +19,337 @@ const Menus = {
         id: `separator-${separatorCounter}`,
         type: "separator",
         contexts,
-        parentId: "save-in-_-_-root"
+        parentId: Menus.IDS.ROOT
       });
       separatorCounter += 1;
     };
 
     return makeSeparatorInner;
   })(),
+
+  setAccesskey: (str, key) => {
+    // CHROME ONLY, FF does not support this yet
+    // https://bugzilla.mozilla.org/show_bug.cgi?id=1320462
+    if (CURRENT_BROWSER !== BROWSERS.CHROME || !key) {
+      return str;
+    }
+
+    if (str.includes(key)) {
+      return str.replace(key, `&${key}`);
+    } else {
+      return `${str} (&${key})`;
+    }
+  },
+
+  addRoot: contexts => {
+    browser.contextMenus.create({
+      id: Menus.IDS.ROOT,
+      title: Menus.setAccesskey(
+        browser.i18n.getMessage("contextMenuRoot"),
+        options.keyRoot
+      ),
+      contexts
+    });
+  },
+
+  addRouteExclusive: contexts => {
+    browser.contextMenus.create({
+      id: "save-in-_-_-route-exclusive",
+      title: Menus.setAccesskey(
+        browser.i18n.getMessage("contextMenuExclusive"),
+        options.keyRoot
+      ),
+      contexts
+    });
+  },
+
+  addSelectionType: contexts => {
+    if (contexts.includes("link")) {
+      browser.contextMenus.create({
+        id: "download-context-media-link",
+        title: browser.i18n.getMessage("contextMenuContextMediaOrLink"),
+        enabled: false,
+        contexts: MEDIA_TYPES.concat("link"),
+        parentId: Menus.IDS.ROOT
+      });
+    } else {
+      browser.contextMenus.create({
+        id: "download-context-media",
+        title: browser.i18n.getMessage("contextMenuContextMedia"),
+        enabled: false,
+        contexts: MEDIA_TYPES,
+        parentId: Menus.IDS.ROOT
+      });
+    }
+
+    if (contexts.includes("selection")) {
+      browser.contextMenus.create({
+        id: "download-context-selection",
+        title: browser.i18n.getMessage("contextMenuContextSelection"),
+        enabled: false,
+        contexts: ["selection"],
+        parentId: Menus.IDS.ROOT
+      });
+    }
+
+    if (contexts.includes("page")) {
+      browser.contextMenus.create({
+        id: "download-context-page",
+        title: browser.i18n.getMessage("contextMenuContextPage"),
+        enabled: false,
+        contexts: ["page"],
+        parentId: Menus.IDS.ROOT
+      });
+    }
+  },
+
+  addOptions: contexts => {
+    browser.contextMenus.create({
+      id: "options",
+      title: browser.i18n.getMessage("contextMenuItemOptions"),
+      contexts,
+      parentId: "save-in-_-_-root"
+    });
+
+    browser.contextMenus.onClicked.addListener(info => {
+      if (info.menuItemId === "options") {
+        browser.runtime.openOptionsPage();
+      }
+    });
+  },
+
+  addShowDefaultFolder: contexts => {
+    browser.contextMenus.create({
+      id: "show-default-folder",
+      title: browser.i18n.getMessage("contextMenuShowDefaultFolder"),
+      contexts,
+      parentId: Menus.IDS.ROOT
+    });
+
+    browser.contextMenus.onClicked.addListener(info => {
+      if (info.menuItemId === "show-default-folder") {
+        browser.downloads.showDefaultFolder();
+      }
+    });
+  },
+
+  addLastUsed: contexts => {
+    const lastUsedTitle =
+      lastUsedPath || browser.i18n.getMessage("contextMenuLastUsed");
+    const lastUsedMenuOptions = {
+      id: `save-in-_-_-last-used`,
+      title: Menus.setAccesskey(lastUsedTitle, options.keyLastUsed),
+      enabled: lastUsedPath ? true : false, // eslint-disable-line
+      contexts,
+      parentId: "save-in-_-_-root"
+    };
+
+    // Chrome, FF < 57 crash when icons is supplied
+    // There is no easy way to detect support, so use a try/catch
+    try {
+      browser.contextMenus.create(
+        Object.assign({}, lastUsedMenuOptions, {
+          icons: {
+            "16": "icons/ic_update_black_24px.svg"
+          }
+        })
+      );
+    } catch (e) {
+      browser.contextMenus.create(lastUsedMenuOptions);
+    }
+  },
+
+  addPaths: (pathsArray, contexts) => {
+    const menuItemCounter = { 0: 0 };
+
+    // Create a stack for nested menus
+    let pathsNestingStack = [];
+    let lastDepth = 0;
+
+    pathsArray.forEach(dir => {
+      // HACK
+      if (dir === SPECIAL_DIRS.SEPARATOR) {
+        Menus.makeSeparator(contexts);
+      } else {
+        const tokens = dir.split("//").map(tok => tok.trim());
+        const depthMatch = tokens[0].match(/^(>+)?(.+)/);
+        const parsedDir = depthMatch[2];
+
+        const validation = new Path.Path(parsedDir).validate();
+        if (!validation.valid) {
+          window.optionErrors.paths.push({
+            message: validation.message,
+            error: `${dir}`
+          });
+
+          return;
+        }
+
+        const comment = (tokens[1] || "").trim();
+        const title = `${parsedDir}${comment ? ` // ${comment}` : ""}`;
+
+        const arrows = depthMatch[1] || "";
+        const depth = arrows.length;
+        if (menuItemCounter[depth] != null) {
+          menuItemCounter[depth] += 1;
+        } else {
+          menuItemCounter[depth] = 1;
+        }
+        const id = `save-in-${menuItemCounter[depth]}-${comment}-${parsedDir}`;
+
+        const parentId =
+          (depth > pathsNestingStack.length
+            ? pathsNestingStack[pathsNestingStack.length - 1]
+            : pathsNestingStack[depth - 1]) || Menus.IDS.ROOT;
+
+        if (depth === 0) {
+          pathsNestingStack = [id];
+        } else if (depth === lastDepth) {
+          pathsNestingStack[depth - 1] = id;
+        } else if (depth > lastDepth) {
+          pathsNestingStack.push(id);
+        }
+        lastDepth = depth;
+        pathsNestingStack = pathsNestingStack.slice(0, depth + 1);
+
+        browser.contextMenus.create({
+          id,
+          title: options.enableNumberedItems
+            ? Menus.setAccesskey(title, menuItemCounter[depth])
+            : title,
+          contexts,
+          parentId
+        });
+      }
+    });
+  },
+
+  // TODO: refactor this to handle only paths, add tests
+  addDownloadListener: () => {
+    browser.contextMenus.onClicked.addListener(info => {
+      if (Object.values(Menus.IDS.TABSTRIP).includes(info.menuItemId)) {
+        return;
+      }
+
+      const matchSave = info.menuItemId.match(/save-in-(\d|_)+-(.*?)-(.*)/);
+
+      if (matchSave && matchSave.length === 4) {
+        let menuIndex = matchSave[1];
+        let comment = matchSave[2];
+        const matchedDir = matchSave[3];
+
+        let url;
+        let suggestedFilename = null;
+        let downloadType = DOWNLOAD_TYPES.UNKNOWN;
+
+        const hasLink = options.links && info.linkUrl;
+
+        if (MEDIA_TYPES.includes(info.mediaType)) {
+          downloadType = DOWNLOAD_TYPES.MEDIA;
+          url = info.srcUrl;
+
+          if (hasLink && options.preferLinks) {
+            downloadType = DOWNLOAD_TYPES.LINK;
+            url = info.linkUrl;
+          }
+        } else if (hasLink) {
+          downloadType = DOWNLOAD_TYPES.LINK;
+          url = info.linkUrl;
+        } else if (options.selection && info.selectionText) {
+          downloadType = DOWNLOAD_TYPES.SELECTION;
+          url = Download.makeObjectUrl(info.selectionText);
+          suggestedFilename = `${Path.truncateIfLongerThan(
+            (currentTab && currentTab.title) || info.selectionText,
+            options.truncateLength - 14
+          )}.selection.txt`;
+        } else if (options.page && info.pageUrl) {
+          downloadType = DOWNLOAD_TYPES.PAGE;
+          url = info.pageUrl;
+          const pageTitle = currentTab && currentTab.title;
+          suggestedFilename = pageTitle || info.pageUrl;
+        } else {
+          return;
+        }
+
+        let saveIntoPath;
+
+        if (matchedDir === "route-exclusive") {
+          saveIntoPath = ".";
+        } else if (matchedDir === "last-used") {
+          saveIntoPath = lastUsedPath;
+          comment = window.lastDownloadState.info.comment;
+          menuIndex = window.lastDownloadState.info.menuIndex;
+        } else {
+          saveIntoPath = matchedDir;
+          lastUsedPath = saveIntoPath;
+          const title = comment
+            ? `${lastUsedPath}${comment ? ` // ${comment}` : ""}`
+            : lastUsedPath;
+
+          if (options.enableLastLocation) {
+            browser.contextMenus.update("save-in-_-_-last-used", {
+              title:
+                CURRENT_BROWSER === BROWSERS.CHROME ? `${title} (&a)` : title,
+              enabled: true
+            });
+          }
+        }
+
+        const parsedPath = new Path.Path(saveIntoPath);
+
+        const saveAsShortcut =
+          (downloadType === DOWNLOAD_TYPES.MEDIA && options.shortcutMedia) ||
+          (downloadType === DOWNLOAD_TYPES.LINK && options.shortcutLink) ||
+          (downloadType === DOWNLOAD_TYPES.PAGE && options.shortcutPage);
+
+        if (saveAsShortcut) {
+          url = Shortcut.makeShortcut(options.shortcutType, url);
+
+          suggestedFilename = Shortcut.suggestShortcutFilename(
+            options.shortcutType,
+            downloadType,
+            info,
+            suggestedFilename,
+            options.truncateLength
+          );
+        }
+
+        if (suggestedFilename) {
+          suggestedFilename = Path.sanitizeFilename(
+            suggestedFilename,
+            options.truncateLength
+          );
+        }
+
+        // Organise things by flattening the info struct and only keeping needed info
+        const opts = {
+          currentTab, // Global
+          linkText: info.linkText,
+          now: new Date(),
+          pageUrl: info.pageUrl,
+          selectionText: info.selectionText,
+          sourceUrl: info.srcUrl || info.url,
+          url, // Changes based off context
+          suggestedFilename, // wip: rename
+          context: downloadType,
+          menuIndex,
+          comment,
+          modifiers: info.modifiers,
+          legacyDownloadInfo: info // wip, remove
+        };
+
+        // keeps track of state of the final path
+        const state = {
+          path: parsedPath,
+          scratch: {},
+          info: opts
+        };
+
+        requestedDownloadFlag = true; // Notifications.
+        Download.renameAndDownload(state);
+      }
+    });
+  },
 
   addTabMenus: () => {
     if (!options.tabEnabled) {